authentication completed

+++ backend/MadisonTime/mt/views.py
@@ -3,13 +3,16 @@ from django.views.generic import (
   ListView, 
   DetailView, 
   CreateView, 
-  UpdateView,
-  DeleteView)
+  UpdateView)
 from django.views.generic.edit import FormMixin
 from django.urls import reverse
+from django.core.exceptions import PermissionDenied
+from braces.views import LoginRequiredMixin, UserPassesTestMixin
+from allauth.account.views import PasswordChangeView
+from allauth.account.models import EmailAddress
 from mt.models import Post, Comment
 from mt.forms import PostForm, CommentForm
-
+from mt.functions import confirmation_required_redirect
 
 # Create your views here.
 def home(request):
@@ -22,12 +25,28 @@ def timetable(request):
   return render(request, 'mt/timetable.html')
 
 def delete_post(request, post_id):
+
+  if not request.user.is_authenticated:
+    raise PermissionDenied
+
   post = get_object_or_404(Post, id=post_id)
+
+  if request.user != post.author:
+    raise PermissionDenied
+
   post.delete()
   return redirect('home')
 
 def delete_comment(request, comment_id):
+
+  if not request.user.is_authenticated:
+    raise PermissionDenied
+
   comment = get_object_or_404(Comment, id=comment_id)
+
+  if request.user != comment.author:
+    raise PermissionDenied
+  
   post_id = comment.post.id
   comment.delete()
   return redirect('post-detail', post_id=post_id)
@@ -45,9 +64,13 @@ class PostDetailView(FormMixin, DetailView):
   form_class = CommentForm
   template_name = "mt/post_detail.html"
 
+  # def test_func(self, user):
+  #   return EmailAddress.objects.filter(user=user, verified=True).exists()
+  
   def get_success_url(self):
     return reverse("post-detail", kwargs={"post_id":self.object.id})
   
+  # FormMixin
   def get_context_data(self, **kwargs):
     context = super(PostDetailView, self).get_context_data(**kwargs)
     context['form'] = CommentForm(initial={'post':self.object})
@@ -56,7 +79,14 @@ class PostDetailView(FormMixin, DetailView):
 
     if comment_id:
       try:
+        if not self.request.user.is_authenticated:
+          raise PermissionDenied
+
         comment_to_edit = Comment.objects.get(id=comment_id, post=self.object)
+
+        if self.request.user != comment_to_edit.author:
+          raise PermissionDenied
+
         context['edit_form'] = CommentForm(instance=comment_to_edit)
         context['edit_comment'] = comment_to_edit
       except Comment.DoesNotExist:
@@ -70,6 +100,13 @@ class PostDetailView(FormMixin, DetailView):
 
   def post(self, request, *args, **kwargs):
     self.object = self.get_object()
+
+    # access control for comments
+    if not request.user.is_authenticated:
+      return redirect(f"{reverse('account_login')}?next={self.request.path}")
+    
+    if not EmailAddress.objects.filter(user=request.user, verified=True).exists():
+      return confirmation_required_redirect(request)
 
     # view identifies the request as an edit request
     if 'edit_comment_id' in request.POST:
@@ -90,11 +127,16 @@ class PostDetailView(FormMixin, DetailView):
       form.instance.post = self.get_object()
     form.save()
     return super(PostDetailView, self).form_valid(form)
+  
 
-class PostCreateView(CreateView):
+class PostCreateView(LoginRequiredMixin, UserPassesTestMixin, CreateView):
   model = Post
   form_class = PostForm
   template_name = "mt/post_form.html"
+
+  # UserPassesTestMixin
+  redirect_unauthenticated_users = True
+  raise_exception = confirmation_required_redirect
 
   def form_valid(self, form):
     form.instance.author = self.request.user
@@ -103,11 +145,24 @@ class PostCreateView(CreateView):
   def get_success_url(self):
     return reverse("post-detail", kwargs={"post_id":self.object.id})
   
-class PostUpdateView(UpdateView):
+  def test_func(self, user):
+    return EmailAddress.objects.filter(user=user, verified=True).exists()
+  
+class PostUpdateView(LoginRequiredMixin, UserPassesTestMixin, UpdateView):
   model = Post
   form_class = PostForm
   template_name = "mt/post_form.html"
   pk_url_kwarg = 'post_id'
 
+  raise_exception = True
+
   def get_success_url(self):
     return reverse("post-detail", kwargs={"post_id":self.object.id})
+  
+  def test_func(self, user):
+    post = self.get_object()
+    return post.author == user
+
+class CustomPasswordChangeView(PasswordChangeView):
+  def get_success_url(self):
+    return reverse("home")
